汇总信息模块: 微小调整，还缺"加工单的统计"

+++ alerp-spring-boot/src/main/java/edu/nju/alerp/service/impl/ShippingOrderServiceImpl.java
@@ -224,7 +224,7 @@ public class ShippingOrderServiceImpl implements ShippingOrderService {
     }
 
     @Override
-    public Integer findTotalInCome(String startDate, String endDate) {
+    public Double findTotalInCome(String startDate, String endDate) {
         return shippingOrderRepository.findTotalInCome(startDate, endDate);
     }
